Added support for different data CSVs
Ting seems to be inconsistent with their CSV output format, at least for
their Data CSVs, sometimes including a Nickname column, and sometimes
not. I added a conditional to sum the correct column given the CSV is in
the format of the two types of Data CSVs I have been notified of.